DPoS vote simple test passed.

+++ Core/Lyra.Core/Authorizers/Vote.cs
@@ -7,7 +7,7 @@ namespace Core.Authorizers
     public class Vote
     {
         public string AccountId { get; set; }
-        public long Amount { get; set; }    // all vote round to int to avoid float calculation 
+        public Decimal Amount { get; set; }    // all vote round to int to avoid float calculation 
         public long ConsHeight { get; set; }
     }
 }
